requirements.txt
pyqt6
mss
pywin32
ollama
pillow
